begin Corpus move to models.rs
