+++ src/data.rs
@@ -7,7 +7,6 @@
 
 //! Data structures and traits for each framework component in the Task store
 
-use rustc_serialize::json::{Json, ToJson};
 use std::collections::{BTreeMap, HashMap};
 use std::fs::File;
 use std::path::Path;
@@ -18,140 +17,6 @@ use postgres::rows::Row;
 use postgres::error::Error;
 
 use Archive::*;
-
-/// A minimalistic ORM trait for `CorTeX` data items
-pub trait CortexORM {
-  /// Select from Task store via the primary id
-  fn select_by_id<'a>(&'a self, connection: &'a Connection) -> Result<Option<Self>, Error>
-  where
-    Self: Sized;
-  /// Select from Task store via a struct-specific uniquely identifying key
-  fn select_by_key<'a>(&'a self, connection: &'a Connection) -> Result<Option<Self>, Error>
-  where
-    Self: Sized;
-  /// Inser the row identified by this struct into the Task store (overwrite if present)
-  fn insert(&self, connection: &Connection) -> Result<(), Error>;
-  /// Delete the row identified by this struct from the Task store
-  fn delete(&self, connection: &Connection) -> Result<(), Error>;
-  /// Construct a struct from a given Task store row
-  fn from_row(row: Row) -> Self;
-  /// Obtain the id of the struct, if any
-  fn get_id(&self) -> Option<i32>;
-}
-
-
-// Tasks
-use std::fmt;
-use super::schema::tasks;
-
-#[derive(Insertable, Clone)]
-#[table_name = "tasks"]
-/// Struct representation of a task store row
-pub struct Task {
-  /// optional id (None for mock / yet-to-be-inserted rows)
-  pub id: Option<i64>,
-  /// id of the service owning this task
-  pub serviceid: i32,
-  /// id of the corpus hosting this task
-  pub corpusid: i32,
-  /// current processing status of this task
-  pub status: i32,
-  /// entry path on the file system
-  pub entry: String,
-}
-impl fmt::Display for Task {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-    // The `f` value implements the `Write` trait, which is what the
-    // write! macro is expecting. Note that this formatting ignores the
-    // various flags provided to format strings.
-    let taskid = match self.id {
-      Some(taskid) => taskid.to_string(),
-      None => "None".to_string(),
-    };
-    write!(
-      f,
-      "(taskid: {}, entry: {},\n\tserviceid: {},\n\tcorpusid: {},\n\t status: {})\n",
-      taskid,
-      self.entry,
-      self.serviceid,
-      self.corpusid,
-      self.status
-    )
-  }
-}
-impl fmt::Debug for Task {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-    // The `f` value implements the `Write` trait, which is what the
-    // write! macro is expecting. Note that this formatting ignores the
-    // various flags provided to format strings.
-    let taskid = match self.id {
-      Some(taskid) => taskid.to_string(),
-      None => "None".to_string(),
-    };
-    write!(
-      f,
-      "(taskid: {},\n\tentry: {},\n\tserviceid: {},\n\tcorpusid: {},\n\t status: {})\n",
-      taskid,
-      self.entry,
-      self.serviceid,
-      self.corpusid,
-      self.status
-    )
-  }
-}
-impl CortexORM for Task {
-  fn get_id(&self) -> Option<i32> {
-    // TODO: Best way to deal with this?
-    match self.id {
-      Some(value) => Some(value as i32),
-      None => None,
-    }
-  }
-  fn select_by_id<'a>(&'a self, connection: &'a Connection) -> Result<Option<Task>, Error> {
-    let stmt = try!(connection.prepare(
-      "SELECT taskid,entry,serviceid,corpusid,status FROM tasks WHERE taskid = $1",
-    ));
-    let rows = try!(stmt.query(&[&self.id]));
-    if rows.len() > 0 {
-      let row = rows.get(0);
-      Ok(Some(Task::from_row(row)))
-    } else {
-      Ok(None)
-    }
-  }
-  fn select_by_key<'a>(&'a self, connection: &'a Connection) -> Result<Option<Task>, Error> {
-    let stmt = try!(connection.prepare("SELECT taskid,entry,serviceid,corpusid,status FROM tasks WHERE entry = $1 and serviceid = $2 and corpusid = $3"));
-    let rows = try!(stmt.query(&[&self.entry, &self.serviceid, &self.corpusid]));
-    if rows.len() > 0 {
-      let row = rows.get(0);
-      Ok(Some(Task::from_row(row)))
-    } else {
-      Ok(None)
-    }
-  }
-  fn insert(&self, connection: &Connection) -> Result<(), Error> {
-    try!(connection.execute("INSERT INTO tasks (serviceid, corpusid, status, entry) values($1, $2, $3, $4) ON CONFLICT(entry, serviceid, corpusid) DO UPDATE SET status=excluded.status;",
-                            &[&self.serviceid, &self.corpusid, &self.status, &self.entry]));
-    Ok(())
-  }
-  fn delete(&self, connection: &Connection) -> Result<(), Error> {
-    try!(connection.execute(
-      "DELETE FROM tasks WHERE taskid = $1",
-      &[&self.id],
-    ));
-    Ok(())
-  }
-  fn from_row(row: Row) -> Self {
-    let fix_width_entry: String = row.get(1);
-    Task {
-      id: Some(row.get(0)),
-      entry: fix_width_entry.trim_right().to_string(),
-      serviceid: row.get(2),
-      corpusid: row.get(3),
-      status: row.get(4),
-    }
-  }
-}
 
 impl Task {
   /// Generates a `TaskReport`, given the path to a result archive from a `CorTeX` processing job
@@ -242,96 +107,6 @@ impl Task {
   }
 }
 
-#[derive(RustcDecodable, RustcEncodable, Clone, Debug)]
-/// A minimal description of a document collection. Defined by a name, path and simple/complex file system setup.
-pub struct Corpus {
-  /// optional id (None for mock / yet-to-be-inserted rows)
-  pub id: Option<i32>,
-  /// a human-readable name for this corpus
-  pub name: String,
-  /// file system path to corpus root
-  /// (a corpus is held in a single top-level directory)
-  pub path: String,
-  /// are we using multiple files to represent a document entry?
-  /// (if unsure, always use "true")
-  pub complex: bool,
-}
-impl Default for Corpus {
-  fn default() -> Self {
-    Corpus {
-      id: None,
-      name: "mock corpus".to_string(),
-      path: ".".to_string(),
-      complex: true,
-    }
-  }
-}
-impl ToJson for Corpus {
-  fn to_json(&self) -> Json {
-    let mut map = BTreeMap::new();
-    map.insert("id".to_string(), self.id.to_json());
-    map.insert("path".to_string(), self.path.to_json());
-    map.insert("name".to_string(), self.name.to_json());
-    map.insert("complex".to_string(), self.complex.to_json());
-    Json::Object(map)
-  }
-}
-
-impl CortexORM for Corpus {
-  fn get_id(&self) -> Option<i32> {
-    self.id
-  }
-  fn select_by_id<'a>(&'a self, connection: &'a Connection) -> Result<Option<Corpus>, Error> {
-    let stmt = try!(connection.prepare(
-      "SELECT corpusid,name,path,complex FROM corpora WHERE corpusid = $1",
-    ));
-    let rows = try!(stmt.query(&[&self.id]));
-    if rows.len() > 0 {
-      let row = rows.get(0);
-      Ok(Some(Corpus::from_row(row)))
-    } else {
-      Ok(None)
-    }
-  }
-  fn select_by_key<'a>(&'a self, connection: &'a Connection) -> Result<Option<Corpus>, Error> {
-    let stmt = try!(connection.prepare(
-      "SELECT corpusid,name,path,complex FROM corpora WHERE name = $1",
-    ));
-    let rows = try!(stmt.query(&[&self.name]));
-    if rows.len() > 0 {
-      let row = rows.get(0);
-      Ok(Some(Corpus::from_row(row)))
-    } else {
-      Ok(None)
-    }
-  }
-  fn insert(&self, connection: &Connection) -> Result<(), Error> {
-    try!(connection.execute(
-      "INSERT INTO corpora (name, path, complex) values($1, $2, $3)",
-      &[&self.name, &self.path, &self.complex],
-    ));
-    Ok(())
-  }
-  fn delete(&self, connection: &Connection) -> Result<(), Error> {
-    try!(connection.execute(
-      "DELETE FROM tasks WHERE corpusid = $1",
-      &[&self.id],
-    ));
-    try!(connection.execute(
-      "DELETE FROM corpora WHERE corpusid = $1",
-      &[&self.id],
-    ));
-    Ok(())
-  }
-  fn from_row(row: Row) -> Self {
-    Corpus {
-      id: Some(row.get(0)),
-      name: row.get(1),
-      path: row.get(2),
-      complex: row.get(3),
-    }
-  }
-}
 impl Corpus {
   /// Return a vector of services currently activated on this corpus
   pub fn select_services<'a>(&'a self, connection: &'a Connection) -> Result<Vec<Service>, Error> {
